refactor: remove adapter folder
