+++ src/main/java/io/pakland/mdas/githubstats/infrastructure/rest/repository/OrganizationGitHubRepository.java
@@ -1,31 +1,30 @@
-package io.pakland.mdas.githubstats.infrastructure.rest.repository.adapters;
+package io.pakland.mdas.githubstats.infrastructure.rest.repository;
 
-import io.pakland.mdas.githubstats.application.dto.RepositoryDTO;
+import io.pakland.mdas.githubstats.application.dto.OrganizationDTO;
 import io.pakland.mdas.githubstats.application.exceptions.HttpException;
-import io.pakland.mdas.githubstats.infrastructure.rest.repository.WebClientConfiguration;
-import io.pakland.mdas.githubstats.domain.repository.RepositoryExternalRepository;
-import java.util.List;
+import io.pakland.mdas.githubstats.domain.repository.OrganizationExternalRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.reactive.function.client.WebClientResponseException;
 
-public class RepositoryGitHubRepository implements RepositoryExternalRepository {
+import java.util.List;
+
+public class OrganizationGitHubRepository implements OrganizationExternalRepository {
 
     private final WebClientConfiguration webClientConfiguration;
     Logger logger = LoggerFactory.getLogger(OrganizationGitHubRepository.class);
 
-    public RepositoryGitHubRepository(WebClientConfiguration webClientConfiguration) {
+    public OrganizationGitHubRepository(WebClientConfiguration webClientConfiguration) {
         this.webClientConfiguration = webClientConfiguration;
     }
 
     @Override
-    public List<RepositoryDTO> fetchTeamRepositories(Integer orgId, Integer teamId)
-        throws HttpException {
+    public List<OrganizationDTO> fetchAvailableOrganizations() throws HttpException {
         try {
             return this.webClientConfiguration.getWebClient().get()
-                .uri(String.format("/organizations/%s/team/%d/repos", orgId, teamId))
+                    .uri("/user/orgs")
                     .retrieve()
-                    .bodyToFlux(RepositoryDTO.class)
+                    .bodyToFlux(OrganizationDTO.class)
                     .collectList()
                     .block();
         } catch (WebClientResponseException ex) {
