+++ src/main/java/io/pakland/mdas/githubstats/infrastructure/shell/controller/UserOptionController.java
@@ -15,9 +15,9 @@ import io.pakland.mdas.githubstats.domain.Repository;
 import io.pakland.mdas.githubstats.domain.Team;
 import io.pakland.mdas.githubstats.domain.User;
 import io.pakland.mdas.githubstats.infrastructure.rest.repository.WebClientConfiguration;
-import io.pakland.mdas.githubstats.infrastructure.rest.repository.adapters.OrganizationGitHubRepository;
-import io.pakland.mdas.githubstats.infrastructure.rest.repository.adapters.RepositoryGitHubRepository;
-import io.pakland.mdas.githubstats.infrastructure.rest.repository.adapters.TeamGitHubRepository;
+import io.pakland.mdas.githubstats.infrastructure.rest.repository.OrganizationGitHubRepository;
+import io.pakland.mdas.githubstats.infrastructure.rest.repository.RepositoryGitHubRepository;
+import io.pakland.mdas.githubstats.infrastructure.rest.repository.TeamGitHubRepository;
 import io.pakland.mdas.githubstats.domain.repository.OrganizationExternalRepository;
 import io.pakland.mdas.githubstats.domain.repository.RepositoryExternalRepository;
 import io.pakland.mdas.githubstats.domain.repository.TeamExternalRepository;
